requirements.txt
deprecated==1.2.18
markupsafe==3.0.2
pyyaml==6.0.2
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
banks==2.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.4
botocore==1.38.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==45.0.3
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
emoji==2.14.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
flask==3.1.1
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
griffe==1.7.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.7.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.44
llama-cloud==0.1.23
llama-cloud-services==0.6.28
llama-cpp-python==0.3.9
llama-index==0.12.40
llama-index-agent-openai==0.4.9
llama-index-cli==0.4.2
llama-index-core==0.12.40
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.2
llama-index-llms-openai==0.4.2
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.8
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.28
lxml==5.4.0
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
mypy-extensions==1.1.0
mysql-connector-python==9.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
olefile==0.47
ollama==0.4.8
openai==1.82.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pip==25.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pygments==2.19.1
pyparsing==3.2.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pywin32==307
pyzmq==26.4.0
rapidfuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.11.3
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
striprtf==0.0.26
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.2
typer-slim==0.15.2
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.36.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
tomli==2.0.1

